# src/main/java/com/example/personal_finance_tracker_api/auth/dto/request/RegisterRequest.java
package com.example.personal_finance_tracker_api.auth.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Pattern;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class RegisterRequest{

    @NotNull(message = "Username cannot be null")
    @Pattern(regexp = "^[a-zA-Z0-9]{3,15}$",
            message = "Username must be 3–15 characters and contain only letters and numbers")
    private String username;

    @NotNull(message = "Password cannot be null")
    private String password;

    @NotNull(message = "Email cannot be null")
    @Email()
    private String email;
}